requirements.txt
PyQt6==6.6.1
mysql-connector-python==8.2.0
pandas==2.1.4
requests==2.31.0
matplotlib==3.8.2